requirements.txt
streamlit>=1.36
pypdf>=4.2.0
python-pptx>=0.6.23
docx2txt>=0.8
openai>=1.35.0
pandas>=2.0.0
